Update README.md
Indeed, there are problems with using online platforms to train deep neural networks (DNNs). For example, there are security issues, such as backdoor attacks, which can manipulate the behaviour of the model with specific triggering patterns. To address this, a long series of backdoor detection methods have been proposed, but they are tailored to specific types of backdoor attacks, i.e. they detect several types of backdoor attacks with different sizes and trigger transparencies. To this end, it may be interesting to propose an adaptive perturbation generation framework, which uses a global-to-local strategy to adapt to multiple backdoor triggering patterns by adjusting the region and budget of adversary attacks with adaptive adversary perturbations. In other words, it is a method of detecting multiple types of backdoor attacks, which have different sizes and levels of transparency, using a strategy that adjusts the region and amount of adversary attacks with adaptive adversary perturbations.